feat(STrack): 将 track_id 类型从 int 改为 uint64_t 并调整初始值
将 `STrack` 类中的 `track_id` 类型从 `int` 改为 `uint64_t`，以支持更大的追踪 ID 范围。
- 修改 `STrack.h` 和 `STrack.cpp` 中的 `track_id` 类型为 `uint64_t`。
- 将 `next_id` 函数的返回类型改为 `uint64_t`，并将初始计数值从 0 改为 1。

refactor(BYTETracker): 调整 BYTETracker 类中相关数据结构以适应 track_id 类型变化

- 修改 `BYTETracker::joint_stracks` 函数中的 `exists` 映射类型为 `std::map<uint64_t, int>`。
- 修改 `BYTETracker::sub_stracks` 函数中的 `stracks` 映射类型为 `std::map<uint64_t, STrack>`。

refactor(global_vars): 将全局变量中的 track_id 类型从 int 改为 uint64_t 并调整初始值

- 修改 `Global::ObjectBox` 结构体中的 `track_id` 类型为 `uint64_t`。
- 将 `track_id` 的初始值从 -1 改为 0，以表示没有追踪。

refactor(detect_utils): 调整 detect_utils 命名空间中相关逻辑以适应 track_id 类型变化

- 修改 `GetColorForId` 函数中的 `id_to_color` 计算逻辑，将条件从 `box.track_id >= 0` 改为 `box.track_id > 0`。
- 修改标签文本生成逻辑中的条件检查，将条件从 `box.track_id >= 0` 改为 `box.track_id > 0`。

+++ src/ByteTrack/STrack.h
@@ -2,6 +2,7 @@
 #define STRACK_H
 #pragma once
 
+#include <cstdint>
 #include <opencv2/opencv.hpp>
 #include "kalmanFilter.h"
 
@@ -30,7 +31,7 @@ namespace ByteTrack
         std::vector<float> to_xyah();
         void mark_lost();
         void mark_removed();
-        int next_id();
+        uint64_t next_id();
         int end_frame();
 
         void activate(byte_kalman::KalmanFilter &kalman_filter, int frame_id);
@@ -40,7 +41,7 @@ namespace ByteTrack
 
     public:
         bool is_activated;
-        int track_id;
+        uint64_t track_id;
         int state;
 
         std::vector<float> _tlwh;
